+++ TagCloudVisualization_Tests/TextFilters_Tests.cs
@@ -3,7 +3,7 @@ using FluentAssertions;
 using NUnit.Framework;
 using TagsCloudVisualization.TextFilters;
 
-namespace TagsCloudVisualization.Tests
+namespace TagsCloudVisualization_Tests
 {
     public class TextFilters_Tests
     {
