+++ TagCloudVisualization_Tests/TextReader_Tests.cs
@@ -3,7 +3,7 @@ using FluentAssertions;
 using NUnit.Framework;
 using TagsCloudVisualization.TextReaders;
 
-namespace TagsCloudVisualization.Tests
+namespace TagsCloudVisualization_Tests
 {
     public class TextReader_Tests
     {
